add remote proc firmware check
add remoteproc firmware check and split the remoteproc sysfs into
dedicated file
